test for wishlist controller added

+++ packages/server/src/controllers/wishlist/remove.ts
@@ -1,7 +1,8 @@
 import provideRepos from '@server/trpc/provideRepos'
-import { authenticatedProcedure } from '@server/trpc/authenticatedProcedure'
 import { wishlistRepository } from '@server/repositories/wishlistRepository'
 import { wishlistSchema, type WishlistForMember } from '@server/entities/wishlist'
+import { authenticatedProcedure } from '@server/auth/aunthenticatedProcedure'
+import { TRPCError } from '@trpc/server'
 
 export default authenticatedProcedure
   .use(
@@ -15,7 +16,23 @@ export default authenticatedProcedure
     })
   )
   .mutation(
-    async ({ input: { id }, ctx: { repos } }): Promise<WishlistForMember> => {
+    async ({ input: { id }, ctx: { repos, authUser } }): Promise<WishlistForMember> => {
+      const existingWishlist = await repos.wishlistRepository.findByEventAndUserId(id, authUser.id)
+      
+      if (!existingWishlist) {
+        throw new TRPCError({
+          code: 'NOT_FOUND',
+          message: 'Wishlist item not found',
+        })
+      }
+
+      if (existingWishlist.userId !== authUser.id) {
+        throw new TRPCError({
+          code: 'FORBIDDEN',
+          message: 'Not authorized to remove this wishlist item',
+        })
+      }
+
       const wishlist = await repos.wishlistRepository.remove(id)
       return wishlist
     }
